Added parameter for the ontology namespace
The concepts returned from the ontology used by the filter are fully
qualified, i.e. they include the namespace, e.g.
http://purl.obolibrary.org/obo/GO_0001234. It is possible that the
annotations in the UIMA pipeline do not include a namespace, e.g. the
annotations in the CRAFT corpus. This new parameter allows a namespace
to be added prior to determining which annotations should be filtered
by comparing the annotations with concepts from the ontology.